pyproject.toml
[tool.poetry]
name = "cardetect"
version = "0.1.0"
description = ""
authors = ["d0wnpour <ratialania01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
boto3 = "^1.26.141"
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
